Refactor tests and use different API and frontend URLs
